refactored code, moved code into container for notepad display
