+++ src/components/NotepadLandingPage/NotepadLandingPage.jsx
@@ -27,12 +27,7 @@ let noteToAdd =
     description: description
 }
     
-// Page Load Functions:
-    useEffect(()=> {
-        dispatch({
-        type: 'FETCH_NOTES'
-        });
-    }, [])
+
 
 // Button Functions 
     // function to send noteToAdd to saga 
@@ -48,22 +43,12 @@ let noteToAdd =
         setHeader('')
 
     }
-    //function to delete note:
-    function handleDelete(id){
-            dispatch({
-                type: 'DELETE_NOTE',
-                payload: id
-            })
-    }
-    //function to send user to detail page where they will edit their note entry
-    function handleEdit(note){  
-      history.push(`/${note.id}`)
-    }
+
  
 
   return (
     <>
-    <NotepadContainerBox />
+    
     <div className="formContainer">
         <h3> Start an Entry:</h3>
     <div className= "entryInput">
@@ -114,30 +99,8 @@ let noteToAdd =
         </button>
 
     </div>
+      <NotepadContainerBox />
 
-{/* Div to store previous notepad entries */}
-    <div>
-        <h1> Notepad Entry History:</h1>
-        <div key={noteEntry.id}>
-    {noteEntry.map(note => {
-            return(
-              <div key={note.id} >
-              <div className= "noteHistoryBox">
-                  <div className="noteHistoryLine"> <h2><u>Date:</u></h2> <h3>{note.date}</h3> </div>
-                 <div className="noteHistoryLine"> <h2><u>Header:</u></h2><h2><b>{note.header}</b></h2></div>
-                 <div className="noteHistoryLine"> <h2><u>Note:</u></h2><h3>{note.description}</h3></div>
-                 <button onClick={() => handleDelete(note.id)}>Delete</button>
-                 <button onClick= {() => handleEdit(note)}> Edit </button>
-              </div>
-             
-            </div>
-            )
-            })}
-  </div>
-
-    {/* button styling is handled in the userPage.css  */}
-    <Link to="/user"><button className="notepadBtn"> <h2>User Homepage</h2></button></Link>
-</div>
     </>
   )
 }
